requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0

# UI Audit System Dependencies
playwright==1.49.0
pytest-playwright==0.5.1